[CHORE] 소스코드 변경 사항 및 패키징 추가 반영
